chore: rename deprecated component PageSettingsModalV1
